requirements.txt
numpy
mss
opencv-python
python-dotenv